Refactor ButtonHandler to ButtonDisplay.
It now extends the Group class of LibGDX, allowing us to pack both
buttons inside it and add it as a whole to the stage.

No tests yet, as we have to figure out how to properly test this. We have
tested it visually, so we are pretty sure it works correctly.