Add missing super call in dispose and extended UserComputerScreenTest
The dispose method of UserComputerScreen was not calling its superclass
dispose method, while it should definitely do that as otherwise the
Stage won't get disposed, possibly slowing down the game.

It turned out that the coverage % of UserComputerScreen was way too low.
Further inspection showed that it could easily be increased.