Added new assets for the font redesign
TTF added for both regular Tahoma and Tahoma bold.
All .fnt and .png files in the main/resources/fonts folder removed, as
they are not used anymore.
Skin file modified to use the new fonts

settings button reduced in size. New .atlas and .png generated for
buttons to account for this change.