Redesigned Solvers
GridSolver now implements Runnable instead of extending TimerTask. This
is done because TimerTask is pretty old, while Executors are relatively
new. Furthermore, and perhaps more importantly, it is now possible to
restart the GridSolver, instead of having to create new objects all the
time.

All classes in the ai package have been modified to be Singletons.
This is done to be able to access the solver from other classes without
having to provide it as a parameter. Bugs that can be resolved by this
are the following:
	HintButton creates a new GridSolver while another is run by
SolveButton.