Implement an annoying tile spawner.
To do this, I have created a new interface `Spawner`, which all new spawners
should implement. I have extracted the random spawn behaviour from Grid into the
RandomSpawner.

The spawner can be chosen via the preferences, however there is not yet a
setting for this. The chosen spawner will be read from here and enabled on the
fly, so you don't have to change anything.

The RandomSpawner is still always used to initiage a Grid, and will remain the
default (meaning that if a grid should have another spawner, you need to
explicitly set this via Grid.setSpawner).