Implemented a validator that validates made moves.
The MoveValidator can check whether a move made was actually valid. This
is done by looking at the current grid and comparing it to the possible
moves from the previous set of tiles. Doing this from the start we can
see if a player has cheated during a game.

If the program find an irregularity in the moves it will notify the
player and his opponent with an overlay.